Improve and rework the buildings and more generlization
