+++ src/main/java/com/minecolonies/coremod/colony/buildings/BuildingCook.java
@@ -8,27 +8,18 @@ import com.minecolonies.coremod.colony.Colony;
 import com.minecolonies.coremod.colony.ColonyView;
 import com.minecolonies.coremod.colony.jobs.AbstractJob;
 import com.minecolonies.coremod.colony.jobs.JobCook;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockFurnace;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.nbt.NBTTagList;
-import net.minecraft.nbt.NBTUtil;
 import net.minecraft.tileentity.TileEntityFurnace;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-import net.minecraftforge.common.util.Constants;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.ArrayList;
-import java.util.List;
-
+import static com.minecolonies.api.util.constant.Constants.STACKSIZE;
 import static com.minecolonies.api.util.constant.Suppression.OVERRIDE_EQUALS;
 
 /**
  * Class of the cook building.
  */
 @SuppressWarnings(OVERRIDE_EQUALS)
-public class BuildingCook extends AbstractBuildingWorker
+public class BuildingCook extends AbstractBuildingFurnaceUser
 {
     /**
      * The cook string.
@@ -41,21 +32,6 @@ public class BuildingCook extends AbstractBuildingWorker
     private static final int MAX_BUILDING_LEVEL = 5;
 
     /**
-     * Tag to store the furnace position.
-     */
-    private static final String TAG_POS = "pos";
-
-    /**
-     * Tag to store the furnace list.
-     */
-    private static final String TAG_FURNACES = "furnaces";
-
-    /**
-     * List of registered furnaces.
-     */
-    private final List<BlockPos> furnaces = new ArrayList<>();
-
-    /**
      * Instantiates a new cook building.
      *
      * @param c the colony.
@@ -64,9 +40,9 @@ public class BuildingCook extends AbstractBuildingWorker
     public BuildingCook(final Colony c, final BlockPos l)
     {
         super(c, l);
-        keepX.put(ItemStackUtils.ISFOOD, Integer.MAX_VALUE);
-        keepX.put(ItemStackUtils.ISCOOKABLE, Integer.MAX_VALUE);
-        keepX.put(TileEntityFurnace::isItemFuel, Integer.MAX_VALUE);
+        keepX.put(ItemStackUtils.ISFOOD, STACKSIZE);
+        keepX.put(ItemStackUtils.ISCOOKABLE, STACKSIZE);
+        keepX.put(TileEntityFurnace::isItemFuel, STACKSIZE);
     }
 
     @NotNull
@@ -94,52 +70,6 @@ public class BuildingCook extends AbstractBuildingWorker
     public String getJobName()
     {
         return COOK_DESC;
-    }
-
-    /**
-     * Return a list of furnaces assigned to this hut.
-     *
-     * @return copy of the list
-     */
-    public List<BlockPos> getFurnaces()
-    {
-        return new ArrayList<>(furnaces);
-    }
-
-    @Override
-    public void writeToNBT(@NotNull final NBTTagCompound compound)
-    {
-        super.writeToNBT(compound);
-        @NotNull final NBTTagList furnacesTagList = new NBTTagList();
-        for (@NotNull final BlockPos entry : furnaces)
-        {
-            @NotNull final NBTTagCompound furnaceCompound = new NBTTagCompound();
-            furnaceCompound.setTag(TAG_POS, NBTUtil.createPosTag(entry));
-            furnacesTagList.appendTag(furnaceCompound);
-        }
-        compound.setTag(TAG_FURNACES, furnacesTagList);
-    }
-
-    @Override
-    public void readFromNBT(@NotNull final NBTTagCompound compound)
-    {
-        super.readFromNBT(compound);
-        final NBTTagList furnaceTagList = compound.getTagList(TAG_FURNACES, Constants.NBT.TAG_COMPOUND);
-        for (int i = 0; i < furnaceTagList.tagCount(); ++i)
-        {
-            furnaces.add(NBTUtil.getPosFromTag(furnaceTagList.getCompoundTagAt(i).getCompoundTag(TAG_POS)));
-        }
-    }
-
-    @Override
-    public void registerBlockPosition(@NotNull final Block block, @NotNull final BlockPos pos, @NotNull final World world)
-    {
-        super.registerBlockPosition(block, pos, world);
-        if (block instanceof BlockFurnace && !furnaces.contains(pos))
-        {
-            furnaces.add(pos);
-        }
-        markDirty();
     }
 
     /**
