+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/cook/EntityAIWorkCook.java
@@ -1,9 +1,12 @@
 package com.minecolonies.coremod.entity.ai.citizen.cook;
 
+import com.minecolonies.api.colony.requestsystem.requestable.Food;
+import com.minecolonies.api.colony.requestsystem.requestable.IRequestable;
 import com.minecolonies.api.util.InventoryUtils;
 import com.minecolonies.api.util.ItemStackUtils;
 import com.minecolonies.api.util.constant.TranslationConstants;
 import com.minecolonies.coremod.colony.jobs.JobCook;
+import com.minecolonies.coremod.colony.requestable.SmeltableOre;
 import com.minecolonies.coremod.entity.EntityCitizen;
 import com.minecolonies.coremod.entity.ai.basic.AbstractEntityAIUsesFurnace;
 import com.minecolonies.coremod.entity.ai.util.AIState;
@@ -15,10 +18,12 @@ import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraftforge.items.wrapper.InvWrapper;
 import org.jetbrains.annotations.NotNull;
 
+import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
 import static com.minecolonies.api.util.constant.Constants.RESULT_SLOT;
+import static com.minecolonies.api.util.constant.Constants.STACKSIZE;
 import static com.minecolonies.api.util.constant.TranslationConstants.COM_MINECOLONIES_COREMOD_ENTITY_BAKER_NO_FURNACES;
 import static com.minecolonies.api.util.constant.TranslationConstants.HUNGRY_INV_FULL;
 import static com.minecolonies.coremod.entity.ai.util.AIState.*;
@@ -68,7 +73,6 @@ public class EntityAIWorkCook extends AbstractEntityAIUsesFurnace<JobCook>
     {
         super(job);
         super.registerTargets(
-                new AITarget(IDLE, START_WORKING),
                 new AITarget(COOK_SERVE_FOOD_TO_CITIZEN, this::serveFoodToCitizen)
         );
         worker.setSkillModifier(CHARISMA_MULTIPLIER * worker.getCitizenData().getCharisma()
@@ -174,5 +178,11 @@ public class EntityAIWorkCook extends AbstractEntityAIUsesFurnace<JobCook>
             return GATHERING_REQUIRED_MATERIALS;
         }
         return START_WORKING;
+    }
+
+    @Override
+    protected IRequestable getSmeltAbleClass()
+    {
+        return new Food(STACKSIZE);
     }
 }
